init 5.0.0

+++ includes/admin-remove-images.php
@@ -1,20 +1,22 @@
 <?php
 add_action('before_delete_post', 'delete_attachments_with_post');
 function delete_attachments_with_post( $postid ){
-    $options = get_option( DT_PLUGIN_NAME );
+  $_post = get_post($postid);
 
-    if( get_post($postid)->post_type == $options['remove-images'] || $options['remove-images'] == 'all' ){
-      $attachments = get_posts( array(
-        'post_type' => 'attachment',
-        'posts_per_page' => -1,
-        'post_status' => '',
-        'post_parent' => $postid
-        ) );
+  if(is_wp_error($_post)) return;
 
-      if( !$attachments )
-          return;
+  $ri = DevelopersTools::$settings['remove-images'];
+  if( $ri != $_post->post_type && $ri != 'all' ) return;
 
-      foreach( $attachments as $attachment )
-          wp_delete_attachment( $attachment->ID, true );  
-    }
+  $attachments = get_posts( array(
+    'post_type' => 'attachment',
+    'posts_per_page' => -1,
+    'post_status' => '',
+    'post_parent' => $postid
+    ) );
+
+  if( is_wp_error($attachments) || !is_array($attachments) || sizeof($attachments) < 1 ) return;
+
+  foreach( $attachments as $attachment )
+    wp_delete_attachment( $attachment->ID, true );
 }
